move export behavior out of main class
